refactor: absorb netinfo legacy-field helpers into the core_io module
GetNetInfoWithLegacyFields() and GetPlatformPort() are not RPC utilities:
they are JSON writers over netInfo and the legacy platform port fields,
consumed by the ToJson() definitions in evo/core_write.cpp plus two report
builders in rpc/masternode.cpp and rpc/quorums.cpp. After the previous commit
hollowed rpc/evo_util down to just these two templates, keeping a separate
rpc/ header for them was pure overhead.

Define them in evo/core_write.cpp next to their main users, declare them in
core_io.h (the module's header, like the other core_write writers), and add
explicit CDeterministicMNState instantiations for the two out-of-module
consumers. rpc/evo_util.h is deleted; the rpc/evo_util module is gone
entirely. No linter suppression changes: the module took part in no cycle.
