+++ src/core_io.h
@@ -7,6 +7,7 @@
 
 #include <consensus/amount.h>
 
+#include <cstdint>
 #include <string>
 #include <vector>
 
@@ -18,6 +19,7 @@ class CTxUndo;
 class uint256;
 struct CMutableTransaction;
 struct CSpentIndexTxInfo;
+enum class MnType : uint16_t;
 
 class UniValue;
 
@@ -55,5 +57,13 @@ std::string EncodeHexTx(const CTransaction& tx);
 std::string SighashToStr(unsigned char sighash_type);
 void ScriptToUniv(const CScript& script, UniValue& out, bool include_hex = true, bool include_address = false);
 void TxToUniv(const CTransaction& tx, const uint256& block_hash, UniValue& entry, bool include_hex = true, int serialize_flags = 0, const CTxUndo* txundo = nullptr, TxVerbosity verbosity = TxVerbosity::SHOW_DETAILS, const CSpentIndexTxInfo* ptxSpentInfo = nullptr);
+
+// evo/core_write.cpp
+/** Reads network info reporting and appends data from legacy fields if applicable */
+template <typename Obj>
+UniValue GetNetInfoWithLegacyFields(const Obj& obj, const MnType& type);
+/** Returns platform port based on purpose and network info version */
+template <bool is_p2p, typename Obj>
+int32_t GetPlatformPort(const Obj& obj);
 
 #endif // BITCOIN_CORE_IO_H
