+++ src/evo/core_write.cpp
@@ -4,6 +4,7 @@
 
 #include <evo/assetlocktx.h>
 #include <evo/cbtx.h>
+#include <evo/dmn_types.h>
 #include <evo/dmnstate.h>
 #include <evo/mnhftx.h>
 #include <evo/netinfo.h>
@@ -11,17 +12,19 @@
 #include <evo/simplifiedmns.h>
 #include <evo/smldiff.h>
 #include <llmq/commitment.h>
-#include <rpc/evo_util.h>
 #include <util/std23.h>
 
 #include <core_io.h>
 #include <key_io.h>
+#include <netaddress.h>
 #include <script/standard.h>
 #include <util/check.h>
 
 #include <univalue.h>
 
+#include <cstdint>
 #include <string>
+#include <type_traits>
 
 namespace {
 template <typename Obj>
@@ -30,6 +33,66 @@ std::string GetDeprecatedServiceField(const Obj& obj)
     return CHECK_NONFATAL(obj.netInfo)->GetPrimary().ToStringAddrPort();
 }
 } // anonymous namespace
+
+template <typename Obj>
+UniValue GetNetInfoWithLegacyFields(const Obj& obj, const MnType& type)
+{
+    UniValue ret{CHECK_NONFATAL(obj.netInfo)->ToJson()};
+
+    // Nothing to do here if not EvoNode, netInfo is empty or netInfo is capable of natively
+    // storing Platform fields
+    if (type != MnType::Evo || obj.netInfo->IsEmpty() || obj.netInfo->CanStorePlatform()) return ret;
+
+    CNetAddr addr{obj.netInfo->GetPrimary()};
+    ret.pushKV(PurposeToString(NetInfoPurpose::PLATFORM_HTTPS).data(),
+               ArrFromService(CService(addr, obj.platformHTTPPort)));
+
+    if constexpr (!std::is_same_v<std::decay_t<Obj>, CSimplifiedMNListEntry>) {
+        // CSimplifiedMNListEntry doesn't store platformP2PPort
+        ret.pushKV(PurposeToString(NetInfoPurpose::PLATFORM_P2P).data(),
+                   ArrFromService(CService(addr, obj.platformP2PPort)));
+    }
+
+    return ret;
+}
+template UniValue GetNetInfoWithLegacyFields(const CDeterministicMNState& obj, const MnType& type);
+
+template <bool is_p2p, typename Obj>
+int32_t GetPlatformPort(const Obj& obj)
+{
+    // Currently, there is nothing that prevents PLATFORM_{HTTPS,P2P} to be registered to
+    // an addr *different* from the primary addr for CORE_P2P. This breaks the assumptions
+    // under which platform{HTTP,P2P}Port operate under (i.e. all three are hosted with the
+    // same addr).
+    //
+    // TODO: Introduce restrictions that enforce this assumption *until* we remove legacy
+    //       fields for good.
+    //
+    bool is_legacy{!(CHECK_NONFATAL(obj.netInfo)->CanStorePlatform())};
+    if constexpr (is_p2p) {
+        static_assert(!std::is_same_v<std::decay_t<Obj>, CSimplifiedMNListEntry>,
+                      "CSimplifiedMNListEntry doesn't have platformP2PPort");
+        if (is_legacy) {
+            return obj.platformP2PPort;
+        }
+        if (obj.netInfo->IsEmpty()) {
+            return -1; // Blank entry, nothing to report
+        }
+        CHECK_NONFATAL(obj.netInfo->HasEntries(NetInfoPurpose::PLATFORM_P2P));
+        return obj.netInfo->GetEntries(NetInfoPurpose::PLATFORM_P2P)[0].GetPort();
+    } else {
+        if (is_legacy) {
+            return obj.platformHTTPPort;
+        }
+        if (obj.netInfo->IsEmpty()) {
+            return -1; // Blank entry, nothing to report
+        }
+        CHECK_NONFATAL(obj.netInfo->HasEntries(NetInfoPurpose::PLATFORM_HTTPS));
+        return obj.netInfo->GetEntries(NetInfoPurpose::PLATFORM_HTTPS)[0].GetPort();
+    }
+}
+template int32_t GetPlatformPort<true>(const CDeterministicMNState& obj);
+template int32_t GetPlatformPort<false>(const CDeterministicMNState& obj);
 
 UniValue CAssetLockPayload::ToJson() const
 {
