+++ src/rpc/masternode.cpp
@@ -9,7 +9,6 @@
 #include <evo/deterministicmns.h>
 #include <governance/superblock.h>
 #include <masternode/payments.h>
-#include <rpc/evo_util.h>
 #include <rpc/json_help.h>
 
 #include <chainparams.h>
