+++ test/util/data/non-backported.txt
@@ -52,7 +52,6 @@ src/qt/proposalmodel.*
 src/qt/proposalresume.*
 src/rpc/coinjoin.cpp
 src/rpc/evo.cpp
-src/rpc/evo_util.*
 src/rpc/governance.cpp
 src/rpc/json_help.*
 src/rpc/masternode.cpp
